moved setup utilities and updated env file

+++ setup_env.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
-python setup_env_util.py "$@"
-chmod +x setup_env_util.sh
-source setup_env_util.sh
-rm setup_env_util.sh
+python mm_scripting_util/setup_env_util.py "$@"
+if [ -e setup_env_util.sh ]; then 
+    echo "running setup script..."
+    chmod +x setup_env_util.sh;
+    source setup_env_util.sh;
+    rm setup_env_util.sh;
+fi
+echo "done"
